Test script (move and collect coins)
To move and collect the test script of the coin code, you need to copy the script paste into the script of the player to run the debug of the implementation code, which can find the loopholes and problems of the code. There is also video evidence of the testers performing the tests in the folder.